Add initial credit endpoint
Summary:
This patch grew out of proportion so I decided to split it in two parts. This patch contains the first one with the minting of tokens (using Caishen's genesis implementation) and a basic `credit` endpoint simply returning success when the transaction worked.

Since that endpoint is supposed to return the new balance of the address, a lot of new code needs to be added to Kubera AND Caishen and therefore should be done inside another patch. I listed all the changes that need to happen in the `BalanceController` module.

Test Plan:
All the tests should be passing and credo shouldn't show any offenses.

Run both Kubera and Caishen with the same RabbitMQ config (exchange, queue, etc).

To credit a user with some points, seed the database (with the `seeds` file) and follow the instruction there to do a `genesis` of X tokens. You can then run the following `curl` request:

```
curl -H "Accept: application/vnd.omisego.v1+json" \
     -H "Content-Type: application/vnd.omisego.v1+json" \
     -d '{"provider_user_id":"123","symbol":"OMG","amount":10000,"metadata":{}}' \
     http://localhost:4000/user.credit_balance
```

Expected response:  `{"success":true}`

Reviewers: sirn, mederic, unnawut

Reviewed By: sirn, unnawut

Subscribers: jenkins

Maniphest Tasks: T155

Differential Revision: https://phabricator.omisego.io/D95