Ensure secret key is hashed
Summary:
Switch secret key to be hashed with Bcrypt before saving so that it
can be displayed only once on creation and make sure it is unretrievable
later on even by us to limit our liability.

Additionally, the "no-op" operation in case user could not be found
is now deliberately slowed down in order to make user enumeration harder.
Since Bcrypt is designed to be slow, the timing leak by not performing
Bcrypt hash operation can be used to verify existance of user.

This differential also provide an automatic migration of previous
plain-text secret key to secret key hash. This migration is done without
using schema to prevent further changes to schema ended up breaking
migration. This migration is irreversible due to the nature of Bcrypt.

Test Plan:
Good scenario:
1. Run `mix ecto.migrate`
2. Make a request with existing access key and secret key pair.
3. See that everything still working.

No-op scenario:
1. Make a request with invalid access key and random secret key.
2. Notice that it's slow.

Reviewers: unnawut, thibault, robin

Reviewed By: unnawut, thibault

Subscribers: jenkins

Maniphest Tasks: T265

Differential Revision: https://phabricator.omisego.io/D146