Introduce types for balances
Summary:
Implement new balance system.

The idea is to allow balances to be used as primary or secondary balances. Also, using a new field named `name`, specific balances can be made like for example a `burn` one (which is generated by default when creating a new account). Users can also create other secondary balances and use one of their IDs in the credit/debit calls to burn tokens instead of getting them back in the primary balance. Note that if no burn balance is specified in the calls, the primary one will be used and funds will never be "burned".

The `debit`/`credit` endpoints now accept an optional `account_id` parameter allowing a provider to specify which primary/burn balance should be used to get/retrieve the tokens from. By default, if the parameter is not set, the account that created the minted token used for the transaction will be used.

Changing the way balances work required changing a lot of things in all the sub apps, including the transaction/transfer logic and the minting process. The endpoint inputs haven't changed, but some of the tests had to be updated to reflect the new changes.

Note that balances aren't lazy-created anymore and are instead generated when an account or a user is created

This patch requires a wipe of the accounts/users/balances/transfers in Kubera and Transactions in Caishen for simplicity (since we are still in a break and increment fast and not in maintenance mode where we can't break anything).

Test Plan: Tests / Credo

Reviewers: mederic, unnawut, sirn

Reviewed By: mederic, unnawut, sirn

Subscribers: jenkins

Maniphest Tasks: T431

Differential Revision: https://phabricator.omisego.io/D254